requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
python-dotenv==1.0.0
requests==2.31.0
pydantic>=2.5.3
openai>=1.30.0
